Added PDFContext
