+++ frontend/app/(sidebarOnly)/viewer/page.tsx
@@ -1,277 +1,54 @@
 "use client";
-import React, { useState, useEffect, useCallback, useRef } from "react";
 import AnnotationBar from "@/components/AnnotationBar";
-import CustomTip from "@/components/CustomTip";
-import AiTip from "@/components/AiTip";
 import AppSidebar from "@/components/app-side-bar";
-
+import DocumentViewer from "@/components/DocumentViewer";
 import {
   SidebarInset,
   Sidebar,
   SidebarContent,
   SidebarHeader,
-  SidebarFooter,
-  SidebarMenuItem,
-  SidebarMenuButton,
-  SidebarInput,
-  SidebarTrigger,
-  SidebarMenu,
 } from "@/components/ui/sidebar";
+import { PDFProvider } from "@/contexts/PDFContext";
 
-import {
-  AreaHighlight,
-  Highlight,
-  PdfHighlighter,
-  PdfLoader,
-  Popup,
-  Tip,
-} from "react-pdf-highlighter";
-import type {
-  Content,
-  IHighlight,
-  NewHighlight,
-  ScaledPosition,
-} from "react-pdf-highlighter";
 
-import "react-pdf-highlighter/dist/style.css";
-
-const getNextId = () => String(Math.random()).slice(2);
-
-const parseIdFromHash = () =>
-  document.location.hash.slice("#highlight-".length);
-
-const resetHash = () => {
-  document.location.hash = "";
-};
-
-const HighlightPopup = ({
-  comment,
-}: {
-  comment: { title: string; text: string; emoji: string };
-}) =>
-  comment.text ? (
-    <div className="Highlight__popup">
-      {comment.emoji} {comment.title}
-    </div>
-  ) : null;
-
-const PRIMARY_PDF_URL = "https://arxiv.org/pdf/1708.08021";
-const SECONDARY_PDF_URL = "https://arxiv.org/pdf/1604.02480";
-
-function DocumentViewer({ pdfUrl = "https://arxiv.org/pdf/1708.08021" }) {
-  // const searchParams = new URLSearchParams(document.location.search);
-  const initialUrl = PRIMARY_PDF_URL;
-
-  const [url, setUrl] = useState(initialUrl);
-  const [highlights, setHighlights] = useState<Array<IHighlight>>([]);
-
-  const resetHighlights = () => {
-    setHighlights([]);
-  };
-
-  const scrollViewerTo = useRef((highlight: IHighlight) => {});
-
-  const scrollToHighlightFromHash = useCallback(() => {
-    const highlight = getHighlightById(parseIdFromHash());
-    if (highlight) {
-      scrollViewerTo.current(highlight);
-    }
-  }, []);
-
-  useEffect(() => {
-    window.addEventListener("hashchange", scrollToHighlightFromHash, false);
-    return () => {
-      window.removeEventListener(
-        "hashchange",
-        scrollToHighlightFromHash,
-        false
-      );
-    };
-  }, [scrollToHighlightFromHash]);
-
-  const getHighlightById = (id: string) => {
-    return highlights.find((highlight) => highlight.id === id);
-  };
-
-  const addHighlight = (highlight: NewHighlight) => {
-    console.log("Saving highlight", highlight);
-    setHighlights((prevHighlights) => [
-      { ...highlight, id: getNextId() },
-      ...prevHighlights,
-    ]);
-  };
-
-  const updateHighlight = (
-    highlightId: string,
-    position: Partial<ScaledPosition>,
-    content: Partial<Content>
-  ) => {
-    console.log("Updating highlight", highlightId, position, content);
-    setHighlights((prevHighlights) =>
-      prevHighlights.map((h) => {
-        const {
-          id,
-          position: originalPosition,
-          content: originalContent,
-          ...rest
-        } = h;
-        return id === highlightId
-          ? {
-              id,
-              position: { ...originalPosition, ...position },
-              content: { ...originalContent, ...content },
-              ...rest,
-            }
-          : h;
-      })
-    );
-  };
-
+function DocumentViewerPage({ pdfUrl = "https://arxiv.org/pdf/1708.08021" }) {
   return (
-    <div className="flex h-full">
-      <div className="h-full">
-        <Sidebar
-          collapsible="icon"
-          className="overflow-hidden [&>[data-sidebar=sidebar]]:flex-row"
-        >
-          <AppSidebar />
-          <Sidebar collapsible="none" className="w-full h-full">
-            <SidebarHeader className="gap-3.5 border-b p-4">
-              <div className="flex w-full items-center justify-between">
-                <div className="text-base font-medium text-foreground">
-                  {/* TODO: TW - Update with documents name */}
-                  Document Name Here...
+    <PDFProvider initialUrl={pdfUrl}>
+      <div className="flex h-full">
+        <div className="h-full">
+          <Sidebar
+            collapsible="icon"
+            className="overflow-hidden [&>[data-sidebar=sidebar]]:flex-row"
+          >
+            <AppSidebar />
+            <Sidebar collapsible="none" className="w-full h-full">
+              <SidebarHeader className="gap-3.5 border-b p-4">
+                <div className="flex w-full items-center justify-between">
+                  <div className="text-base font-medium text-foreground">
+                    {/* TODO: TW - Update with documents name */}
+                    Document Name Here...
+                  </div>
                 </div>
-              </div>
-              {/* TODO: TW - Can add searchinf of annotations here */}
-              {/* <SidebarInput className="" placeholder="Type to search..." /> */}
-            </SidebarHeader>
-            <SidebarContent className="flex-1">
-              <div className="h-full w-full">
-                <AnnotationBar
-                  highlights={highlights}
-                  resetHighlights={() => setHighlights([])}
-                />
-              </div>
-            </SidebarContent>
+                {/* TODO: TW - Can add searchinf of annotations here */}
+                {/* <SidebarInput className="" placeholder="Type to search..." /> */}
+              </SidebarHeader>
+              <SidebarContent className="flex-1">
+                <div className="h-full w-full">
+                  <AnnotationBar />
+                </div>
+              </SidebarContent>
+            </Sidebar>
           </Sidebar>
-        </Sidebar>
-      </div>
+        </div>
 
-      <div className="flex flex-1 flex-col">
-        <SidebarInset className="">
-          <div className="flex w-full flex-1 flex-col">
-            <div className="h-full relative overflow-auto">
-              <PdfLoader url={pdfUrl} beforeLoad={<div>Loading PDF...</div>}>
-                {(pdfDocument) => (
-                  <>
-                    <div className="absolute inset-0">
-                      <PdfHighlighter
-                        pdfDocument={pdfDocument}
-                        enableAreaSelection={(event) => event.altKey}
-                        onScrollChange={resetHash}
-                        scrollRef={(scrollTo) => {
-                          scrollViewerTo.current = scrollTo;
-                          scrollToHighlightFromHash();
-                        }}
-                        onSelectionFinished={(
-                          position,
-                          content,
-                          hideTipAndSelection,
-                          transformSelection
-                        ) => {
-                          return (
-                            <div className="flex flex-row gap-2">
-                              <CustomTip
-                                onOpen={transformSelection}
-                                onConfirm={(comment) => {
-                                  addHighlight({
-                                    content,
-                                    position,
-                                    comment: {
-                                      title: comment.title || "No Title",
-                                      text: comment.text || "No comment",
-                                      emoji: "💬",
-                                    },
-                                  });
-                                  hideTipAndSelection();
-                                }}
-                              />
-                              {/* <AiTip
-                                onOpen={transformSelection}
-                                onConfirm={(comment) => {
-                                  addHighlight({
-                                    content,
-                                    position,
-                                    comment: {
-                                      title: comment.title || "No Title",
-                                      text: comment.text || "No comment",
-                                      emoji: "💬",
-                                    },
-                                  });
-                                  hideTipAndSelection();
-                                }}
-                              /> */}
-                            </div>
-                          );
-                        }}
-                        highlightTransform={(
-                          highlight,
-                          index,
-                          setTip,
-                          hideTip,
-                          viewportToScaled,
-                          screenshot,
-                          isScrolledTo
-                        ) => {
-                          const isTextHighlight = !highlight.content?.image;
-                          const component = isTextHighlight ? (
-                            <Highlight
-                              isScrolledTo={isScrolledTo}
-                              position={highlight.position}
-                              comment={highlight.comment}
-                            />
-                          ) : (
-                            <AreaHighlight
-                              isScrolledTo={isScrolledTo}
-                              highlight={highlight}
-                              onChange={(boundingRect) => {
-                                updateHighlight(
-                                  highlight.id,
-                                  {
-                                    boundingRect:
-                                      viewportToScaled(boundingRect),
-                                  },
-                                  { image: screenshot(boundingRect) }
-                                );
-                              }}
-                            />
-                          );
-                          return (
-                            <Popup
-                              popupContent={<HighlightPopup {...highlight} />}
-                              onMouseOver={(popupContent) =>
-                                setTip(highlight, (highlight) => popupContent)
-                              }
-                              onMouseOut={hideTip}
-                              key={index}
-                            >
-                              {component}
-                            </Popup>
-                          );
-                        }}
-                        highlights={highlights}
-                      ></PdfHighlighter>
-                    </div>
-                  </>
-                )}
-              </PdfLoader>
-            </div>
-          </div>
-        </SidebarInset>
+        <div className="flex flex-1 flex-col">
+          <SidebarInset className="">
+           <DocumentViewer />
+          </SidebarInset>
+        </div>
       </div>
-    </div>
+    </PDFProvider>
   );
 }
 
-export default DocumentViewer;
+export default DocumentViewerPage;
